// types/cardTypes.ts
export type cardTypes = {
    id_carte: number;// 01 a 10 caracteres numerique
    cartNumber: number;// 16 caracteres numerique
    month: number;// 02 caracteres numerique 
    year: number;// 02 caracteres numerique
    default:boolean;// true or false boolean
    cardIdStripe: string; // id stripe de la carte
}
    